requirements.txt
# Smart Resume Matching Engine — Python Dependencies
# Phase 2: Core dependencies for Link (connectivity verification)

# --- Web Framework ---
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# --- ML / NLP ---
sentence-transformers>=2.3.0
spacy>=3.7.0
numpy>=1.26.0
scipy>=1.12.0

# --- Document Parsing ---
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
python-docx>=1.1.0

# --- Database ---
psycopg2-binary>=2.9.9

# --- Cache ---
redis>=5.0.0

# --- Config ---
python-dotenv>=1.0.0
